requirements.txt
Click==7.0
configparser==4.0.2
dlib==19.19.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.1
Pillow==7.0.0
PyMySQL==0.9.3
Werkzeug==1.0.0
